+++ sailsim/boat/boatgetset.py
@@ -3,12 +3,14 @@ from math import pi
 from sailsim.utils.anglecalculations import angleKeepInterval, directionKeepInterval
 
 
-def setBoat(self, posX, posY, speedX=0, speedY=0):
+def setBoat(self, posX, posY, direction=0, speedX=0, speedY=0, angSpeed=0):
     """Set important properties of boat."""
     self.posX = posX
     self.posY = posY
+    self.direction = direction
     self.speedX = speedX
     self.speedY = speedY
+    self.angSpeed = angSpeed
 
 
 def getPos(self):
@@ -47,7 +49,7 @@ def setRudderAngle(self, rudderAngle):
 
 
 def setRudderAngleDeg(self, rudderAngle):
-    """Convert degrees to radiants and run setMainSailAngle()."""
+    """Convert degrees to radiants and run setRudderAngle()."""
     self.setRudderAngle(rudderAngle * pi / 180)