+++ sailsim/simulation/Frame.py
@@ -9,6 +9,7 @@ class Frame():
         self.boatPosX = self.boatPosY = None
         self.boatSpeedX = self.boatSpeedY = None
         self.boatDirection = None
+        self.boatAngSpeed = None
 
         self.boatMainSailAngle = None
         self.boatRudderAngle = None
@@ -21,8 +22,15 @@ class Frame():
         self.boatForceX = self.boatForceY = None
         self.boatSailDragX = self.boatSailDragY = None
         self.boatSailLiftX = self.boatSailLiftY = None
-        self.boatWaterDragX = self.boatWaterDragY = None
-        self.boatWaterLiftX = self.boatWaterLiftY = None
+        self.boatCenterboardDragX = self.boatCenterboardDragY = None
+        self.boatCenterboardLiftX = self.boatCenterboardLiftY = None
+
+        self.boatRudderDragX = self.boatRudderDragY = None
+        self.boatRudderLiftX = self.boatRudderLiftY = None
+
+        self.boatTorque = None
+        self.boatWaterDragTorque = None
+        self.boatCenterboardTorque = self.boatRudderTorque = None
 
 
     def collectSimulation(self, simulation):
@@ -37,6 +45,7 @@ class Frame():
         self.boatSpeedX = boat.speedX
         self.boatSpeedY = boat.speedY
         self.boatDirection = boat.direction
+        self.boatAngSpeed = boat.angSpeed
 
         self.boatMainSailAngle = boat.mainSailAngle
         self.boatRudderAngle = boat.rudderAngle
@@ -50,8 +59,16 @@ class Frame():
         (self.boatForceX, self.boatForceY) = (h.forceX, h.forceY)
         (self.boatSailDragX, self.boatSailDragY) = (h.sailDragX, h.sailDragY)
         (self.boatSailLiftX, self.boatSailLiftY) = (h.sailLiftX, h.sailLiftY)
-        (self.boatWaterDragX, self.boatWaterDragY) = (h.waterDragX, h.waterDragY)
-        (self.boatWaterLiftX, self.boatWaterLiftY) = (h.waterLiftX, h.waterLiftY)
+        (self.boatCenterboardDragX, self.boatCenterboardDragY) = (h.centerboardDragX, h.centerboardDragY)
+        (self.boatCenterboardLiftX, self.boatCenterboardLiftY) = (h.centerboardLiftX, h.centerboardLiftY)
+
+        (self.boatRudderDragX, self.boatRudderDragY) = (h.rudderDragX, h.rudderDragY)
+        (self.boatRudderLiftX, self.boatRudderLiftY) = (h.rudderLiftX, h.rudderLiftY)
+
+        self.boatTorque = h.torque
+        self.boatWaterDragTorque = h.waterDragTorque
+        self.boatCenterboardTorque = h.centerboardTorque
+        self.boatRudderTorque = h.rudderTorque
 
     def collectWind(self, wind, x, y, size, distance):
         """Collect and save all information about the wind."""
@@ -74,12 +91,14 @@ class Frame():
         """Return string that contains all data about this frame."""
         data = [
             self.frameNr, self.time,
-            self.boatPosX, self.boatPosY, self.boatSpeedX, self.boatSpeedY, self.boatDirection,
+            self.boatPosX, self.boatPosY, self.boatSpeedX, self.boatSpeedY, self.boatDirection, self.boatAngSpeed,
             self.boatMainSailAngle, self.boatRudderAngle,
             self.boatApparentWindX, self.boatApparentWindY, self.boatApparentWindAngle, self.boatLeewayAngle, self.boatAngleOfAttack,
             self.boatForceX, self.boatForceY,
             self.boatSailDragX, self.boatSailDragY, self.boatSailLiftX, self.boatSailLiftY,
-            self.boatWaterDragX, self.boatWaterDragY, self.boatWaterLiftX, self.boatWaterLiftY,
+            self.boatCenterboardDragX, self.boatCenterboardDragY, self.boatCenterboardLiftX, self.boatCenterboardLiftY,
+            self.boatRudderDragX, self.boatRudderDragY, self.boatRudderLiftX, self.boatRudderLiftY,
+            self.boatTorque, self.boatWaterDragTorque, self.boatCenterboardTorque, self.boatRudderTorque,
         ]
         data.extend(self.getWindList())
         dataStr = [f'{x:.4f}'.rstrip('0').rstrip('.') for x in data] # FIXME very slow and inflexible