+++ sailsim/simulation/Simulation.py
@@ -24,13 +24,17 @@ class Simulation:
         self.frame = 0
         self.lastFrame = lastFrame
 
-    def run(self):
+    def run(self, steps=0):
         """Run whole Simulation if lastFrame is set."""
-        # Check if lastFrame exisists
-        if self.lastFrame is None:
-            raise Exception('Simulation has no lastFrame')
-        while self.frame <= self.lastFrame:
-            self.step()
+        if steps < 1:
+            # Check if lastFrame exisists
+            if self.lastFrame is None:
+                raise Exception('Simulation has no lastFrame')
+            while self.frame <= self.lastFrame:
+                self.step()
+        else:
+            for i in range(steps):
+                self.step()
 
     def step(self):
         """Run one step of the Simulation."""
@@ -40,7 +44,7 @@ class Simulation:
         # Calculate Forces on boat
         (boatX, boatY) = self.world.boat.getPos()                           # Fetch boat position
         (windX, windY) = self.world.wind.getWindCart(boatX, boatY, time)    # Get wind
-        (forceX, forceY) = self.world.boat.resultingForce(windX, windY)
+        (forceX, forceY, torque) = self.world.boat.resultingCauses(windX, windY)
 
         # Save frame
         self.frameList.grabFrame(self)
@@ -49,7 +53,7 @@ class Simulation:
         self.world.boat.runSailor()
 
         # Move Boat
-        self.world.boat.applyForce(forceX, forceY, self.timestep)
+        self.world.boat.applyCauses(forceX, forceY, torque, self.timestep)
         self.world.boat.moveInterval(self.timestep)